requirements.txt
flask==2.3.2
requests==2.31.0
numpy==1.25.2
opencv-python==4.8.0.76
google-cloud-vision==3.3.1
gspread==5.11.2
google-auth==2.23.3
gunicorn==20.1.0
python-dotenv==1.0.1
